Add html elements
Added an h2 element with text.
Added a sidebar with background image and logo.
Added a form element with two fieldset inside it.
Added related elements like name, email, etc in the first fieldset.
Added a submit button and some text below it with a
placeholder link in the second fieldset.