Make changes
Tweaked some of the code  and added some more styles.
The site is working mostly fine but need to fix the button background
to remove the light grey color and switch to white background.
Have some ideas on how to fix . Grouping the HTML elements separately
should do the trick.